Added Akhenaten to build server

+++ Akhenaten/macsourceports_universal2.sh
@@ -4,7 +4,7 @@ export PRODUCT_NAME="Akhenaten"
 export PROJECT_NAME="Akhenaten"
 export PORT_NAME="Akhenaten"
 export ICONSFILENAME="akhenaten"
-export EXECUTABLE_NAME="Akhenaten"
+export EXECUTABLE_NAME="akhenaten"
 export GIT_TAG="v1.0"
 export GIT_DEFAULT_BRANCH="master"
 
@@ -13,38 +13,56 @@ source ../common/constants.sh
 
 cd ../../${PROJECT_NAME}
 
-# reset to the main branch
-echo git checkout ${GIT_DEFAULT_BRANCH}
-git checkout ${GIT_DEFAULT_BRANCH}
+if [ "$1" == "buildserver" ] || [ "$2" == "buildserver" ]; then
+	echo "Skipping git because we're on the build server"
+else
+    # reset to the main branch
+    echo git checkout ${GIT_DEFAULT_BRANCH}
+    git checkout ${GIT_DEFAULT_BRANCH}
 
-# fetch the latest 
-echo git pull
-git pull
+    # fetch the latest 
+    echo git pull
+    git pull
 
-# check out the latest release tag
-# echo git checkout tags/${GIT_TAG}
-# git checkout tags/${GIT_TAG}
+	# skipping checkout since we just use the latest on this one
+fi
 
 rm -rf ${BUILT_PRODUCTS_DIR}
 
-# create makefiles with cmake, perform builds with make
-rm -rf ${X86_64_BUILD_FOLDER}
-mkdir ${X86_64_BUILD_FOLDER}
-cd ${X86_64_BUILD_FOLDER}
-/usr/local/bin/cmake ..
-make -j$NCPU
+if [ "$1" == "buildserver" ] || [ "$2" == "buildserver" ]; then
+    mkdir ${BUILT_PRODUCTS_DIR}
+    cd ${BUILT_PRODUCTS_DIR}
+    cmake \
+    -DCMAKE_OSX_ARCHITECTURES="arm64;x86_64" \
+    -DCMAKE_OSX_DEPLOYMENT_TARGET=10.7 \
+    ..
+    cmake --build . --parallel $NCPU
+    install_name_tool -add_rpath @executable_path/. ${EXECUTABLE_FOLDER_PATH}/${EXECUTABLE_NAME}
+    "../../MSPBuildSystem/common/copy_dependencies.sh" ${EXECUTABLE_FOLDER_PATH}/${EXECUTABLE_NAME}
+else
+    # create makefiles with cmake, perform builds with make
+    rm -rf ${X86_64_BUILD_FOLDER}
+    mkdir ${X86_64_BUILD_FOLDER}
+    cd ${X86_64_BUILD_FOLDER}
+    /usr/local/bin/cmake ..
+    make -j$NCPU
 
-cd ..
-rm -rf ${ARM64_BUILD_FOLDER}
-mkdir ${ARM64_BUILD_FOLDER}
-cd ${ARM64_BUILD_FOLDER}
-cmake ..
-make -j$NCPU
+    cd ..
+    rm -rf ${ARM64_BUILD_FOLDER}
+    mkdir ${ARM64_BUILD_FOLDER}
+    cd ${ARM64_BUILD_FOLDER}
+    cmake ..
+    make -j$NCPU
+fi
 
 cd ..
 
 # create the app bundle
-"../MSPBuildSystem/common/build_app_bundle.sh"
+if [ "$1" == "buildserver" ] || [ "$2" == "buildserver" ]; then
+    "../MSPBuildSystem/common/build_app_bundle.sh" "skiplipo" "skiplibs"
+else
+    "../MSPBuildSystem/common/build_app_bundle.sh"
+fi
 
 #sign and notarize
 "../MSPBuildSystem/common/sign_and_notarize.sh" "$1"
